Split out `match_wild_err_arm`
